refactor: shared methods with other commands
